+++ src/Config.php
@@ -6,44 +6,41 @@ use PHPForm\Renderers\TwigRenderer;
 class Config extends Singleton
 {
     /**
-     * Template packs available
-     */
-    const TEMPLATE_PACKS = array("default", "bootstrap4");
-
-    /**
-     * Directory with templates
-     * @var string
+     * @var string Directory with templates
      */
     protected $templates_dir = __DIR__ . '/templates/';
 
     /**
-     * Template pack defined
-     * @var string
+     * @var string Default fallback template pack
      */
-    protected $template_pack = "default";
+    protected $fallback_template_pack = "default";
 
     /**
-     * Renderer class used to render html content
-     * @var string
+     * @var string Default template pack defined. If null, fallback is used.
+     */
+    protected $template_pack = null;
+
+    /**
+     * @var string Renderer class used to render html content
      */
     protected $renderer_class = TwigRenderer::class;
 
     /**
-     * Renderer instance based on $renderer_class.
-     * @var PHPForm\Renderers\Renderer
+     * @var PHPForm\Renderers\Renderer Renderer instance based on $renderer_class.
      */
     private $renderer;
 
     /**
      * Templates path accordingly with defined template pack.
+     *
      * @return string
      */
-    private function getTemplatesPath()
+    private function buildPath($template_pack)
     {
-        $path = $this->templates_dir . $this->template_pack . "/";
+        $path = $this->templates_dir . $template_pack . "/";
 
         if (!file_exists($path)) {
-            trigger_error("Template dir '$path' don't exists.", E_USER_ERROR);
+            trigger_error("Template pack dir '$path' don't exists.", E_USER_ERROR);
         }
 
         return $path;
@@ -51,12 +48,16 @@ class Config extends Singleton
 
     /**
      * Return renderer class instantiated
+     *
      * @return PHPForm\Renderers\Renderer
      */
     public function getRenderer()
     {
         if (is_null($this->renderer)) {
-            $this->renderer = new $this->renderer_class($this->getTemplatesPath());
+            $fallback_pack_dir = $this->buildPath($this->fallback_template_pack);
+            $pack_dir = !is_null($this->template_pack) ? $this->buildPath($this->template_pack) : null;
+
+            $this->renderer = new $this->renderer_class($fallback_pack_dir, $pack_dir);
         }
 
         return $this->renderer;
@@ -64,15 +65,11 @@ class Config extends Singleton
 
     /**
      * Define template pack
+     *
      * @param string
      */
     public function setTemplatePack($template_pack)
     {
-        if (!in_array($template_pack, self::TEMPLATE_PACKS)) {
-            $packs = implode(", ", self::TEMPLATE_PACKS);
-            trigger_error("Template pack '$template_pack' not valid. Available packs are: $packs", E_USER_ERROR);
-        }
-
         $this->template_pack = $template_pack;
     }
 }